Support arbitrary sized watchpoints
In LLDB command line you can specify any variable and it will attempt to
create a series of watchpoints, regardless of the size of the requested
variable or range. We do the same: if a non-native size is requested,
then create N watchpoints of word-size contiguously from the base
address. If any one of them fails, delete all of them.